requirements.txt
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Data handling
pandas>=2.0.0
pillow>=10.0.0
opencv-python>=4.8.0

# Dataset download
kaggle>=1.5.16
opendatasets>=0.1.22

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# Experiment tracking (optional)
tensorboard>=2.13.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# Data Exploration
streamlit>=1.28.0
